Add minimal dependency tracking and rename contexts
Rename RenderContext -> LayoutContext and ComponentContext -> RenderContext
so names reflect actual usage: LayoutContext builds the component tree,
RenderContext is in scope during rendering.

Add signal.get(using RenderContext) to enforce render-scope reads at the
type level, and signal.peek for untracked reads in event handlers and
setup code. Add @implicitNotFound on RenderContext with an actionable
error message. Move isFocused from AppContext to RenderContext so it is
accessible in LeafContent scope.
