+++ ui/shared/src/main/scala/terminus/ui/AppContext.scala
@@ -20,29 +20,22 @@ import terminus.Key
 
 /** Combined capability context for interactive component bodies.
   *
-  * Bundles [[ComponentContext]], [[EventContext]], and [[RenderContext]] into a
+  * Bundles [[RenderContext]], [[EventContext]], and [[LayoutContext]] into a
   * single capability so layout components (Row, Column) can thread all three
   * through their bodies without requiring three separate context parameters.
   *
-  * Layout components create a child AppContext that forwards ComponentContext
-  * and EventContext from the parent while substituting their own RenderContext.
+  * Layout components create a child AppContext that forwards RenderContext and
+  * EventContext from the parent while substituting their own LayoutContext.
   */
-trait AppContext extends ComponentContext, EventContext, RenderContext:
-  /** Returns true if this context is inside a focused [[FocusScope]], or if it
-    * is not inside any [[FocusScope]] at all (global scope is always active).
-    *
-    * Components use this during rendering to adjust their appearance (e.g.
-    * highlight a border) when they have focus.
-    */
-  def isFocused: Boolean = true
+trait AppContext extends RenderContext, EventContext, LayoutContext
 
 /** Content of a leaf component (e.g. Text).
   *
-  * The [[ComponentContext]] in scope enables signal reads to register the
+  * The [[RenderContext]] in scope enables signal reads to register the
   * component as a subscriber, so that signal changes can trigger a targeted
   * re-render.
   */
-type LeafContent[A] = ComponentContext ?=> A
+type LeafContent[A] = RenderContext ?=> A
 
 /** Body of a layout component (e.g. Row, Column) or the top-level app body.
   *
@@ -53,13 +46,13 @@ type AppContent[A] = AppContext ?=> A
 
 object AppContext:
 
-  /** Creates a child AppContext that delegates [[ComponentContext]] and
-    * [[EventContext]] to the parent, but uses rc as the [[RenderContext]].
+  /** Creates a child AppContext that delegates [[RenderContext]] and
+    * [[EventContext]] to the parent, but uses rc as the [[LayoutContext]].
     *
     * Used by layout components (Row, Column) to substitute their own layout
     * context while preserving the reactive context from the parent.
     */
-  def child(parent: AppContext, rc: RenderContext): AppContext =
+  def child(parent: AppContext, rc: LayoutContext): AppContext =
     new AppContext:
       private[ui] def invalidate(): Unit = parent.invalidate()
       def createSignal[A](initial: A): Signal[A] = parent.createSignal(initial)
