+++ ui/shared/src/main/scala/terminus/ui/RenderContext.scala
@@ -16,12 +16,36 @@
 
 package terminus.ui
 
+import scala.annotation.implicitNotFound
+
+/** Reactive scope for a component render pass.
+  *
+  * Signal reads inside a component's content use the RenderContext in scope to
+  * register the component as a subscriber, so that signal changes can trigger a
+  * targeted re-render.
+  *
+  * For the initial full-frame re-render implementation this is a stub; the full
+  * dependency-tracking mechanism will be wired in when subtree re-render is
+  * implemented.
+  */
+@implicitNotFound(
+  """signal.get requires a RenderContext, which is only available inside a
+component content lambda (the body passed to Text, etc.).
+
+If you meant to read this signal reactively, move the call inside a
+component's content: Text(w, h) { mySignal.get }
+
+If you intentionally want an untracked read in setup or event-handler
+code, use signal.peek instead."""
+)
 trait RenderContext:
-  def size: Size
-  def add(component: Component): Unit
+  private[ui] def invalidate(): Unit
 
-trait RootContext extends RenderContext:
-  def render(using Terminal): Unit
-
-trait ChildContext extends RenderContext:
-  def render(bounds: Rect, buf: Buffer): Unit
+  /** Returns true if this component is inside a focused [[FocusScope]], or if
+    * it is not inside any [[FocusScope]] at all.
+    *
+    * Must be read in render scope (inside a component content lambda) to
+    * reflect the current focus state. Reading it in setup scope captures the
+    * value at construction time and will not update.
+    */
+  def isFocused: Boolean = true
